punto partida ejercicio repaso

+++ MvcNetCorePaginacionRegistros/Models/Empleado.cs
@@ -14,7 +14,7 @@ namespace MvcNetCorePaginacionRegistros.Models
         [Column("OFICIO")]
         public string Oficio { get; set; }
         [Column("SALARIO")]
-        public decimal Salario { get; set; }
+        public int Salario { get; set; }
         [Column("DEPT_NO")]
         public int IdDepartamento { get; set; }
     }
